getRemainingBalance function
